requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
openpyxl>=3.1.0

# Search + fetch — no API keys required
ddgs>=9.0.0
httpx>=0.27.0
trafilatura>=1.12.0

# LLM providers — install only what you use
openai>=1.30.0
google-generativeai>=0.7.0
anthropic>=0.28.0
